# Improved tm-monitor formatting